Cleanup and added laghos_amr files
